+++ include/rocksdb/listener.h
@@ -12,6 +12,7 @@
 #include <unordered_map>
 #include <vector>
 #include "rocksdb/compaction_job_stats.h"
+#include "rocksdb/compression_type.h"
 #include "rocksdb/status.h"
 #include "rocksdb/table_properties.h"
 
@@ -24,7 +25,6 @@ class DB;
 class ColumnFamilyHandle;
 class Status;
 struct CompactionJobStats;
-enum CompressionType : unsigned char;
 
 enum class TableFileCreationReason {
   kFlush,
